bot/requirements.txt
requests
locust
pymongo
python-dotenv
